Moved ForecastContext and repository classes from Library project to Context project.
