+++ src/Housing.Forecast.Context/ForecastContext.cs
@@ -10,7 +10,7 @@ using Housing.Forecast.Library;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.AspNetCore.Hosting;
 
-namespace Housing.Forecast.Library
+namespace Housing.Forecast.Context
 {
     /// <summary>
     /// A ForecastContext instance represents a session with the database and is used to query and save instances of the entities.
