+++ src/Housing.Forecast.Context/Repos/RoomRepo.cs
@@ -5,7 +5,7 @@ using System.Text;
 using Housing.Forecast.Library;
 using Housing.Forecast.Library.Models;
 
-namespace Housing.Forecast.Library.Repos
+namespace Housing.Forecast.Context.Repos
 {
     public class RoomRepo : IRepo<Room>, IDisposable
     {
