+++ src/Housing.Forecast.Testing/Context/UserRepoTests.cs
@@ -5,17 +5,17 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 
-using Housing.Forecast.Library;
+using Housing.Forecast.Context;
+using Housing.Forecast.Context.Repos;
 using Housing.Forecast.Library.Models;
-using Housing.Forecast.Library.Repos;
 using Housing.Forecast.Service;
 
 namespace Housing.Forecast.Testing.Library
 {
-    public class RoomRepoTests
+    public class UserRepoTests
     {
         private ForecastContext _context;
-        private IRepo<Room> _roomRepository;
+        private IRepo<User> _userRepository;
         private static DbContextOptions<ForecastContext> options;
 
         [Fact]
@@ -24,16 +24,16 @@ namespace Housing.Forecast.Testing.Library
             init();
             using(_context = new ForecastContext(options)) {
                 // Arrange
-                IEnumerable<Room> rooms;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                IEnumerable<User> users;
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
                 _context.SaveChanges();
 
                 // Act
-                rooms = _roomRepository.Get();
+                users = _userRepository.Get();
 
                 // Assert
-                Assert.NotEmpty(rooms);
+                Assert.NotEmpty(users);
             }
         }
 
@@ -43,17 +43,17 @@ namespace Housing.Forecast.Testing.Library
             init();
             using(_context = new ForecastContext(options)) {
                 // Arrange
-                IEnumerable<Room> rooms;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                IEnumerable<User> users;
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
                 _context.SaveChanges();
-                Room room = _context.Rooms.FirstOrDefault();
+                User user = _context.Users.FirstOrDefault();
 
                 // Act
-                rooms = _roomRepository.GetByDate(room.Created);
+                users = _userRepository.GetByDate(user.Created);
 
                 // Assert
-                Assert.Equal(rooms.FirstOrDefault().Id, room.Id);
+                Assert.Equal(users.FirstOrDefault().Id, user.Id);
             }
         }
 
@@ -62,17 +62,17 @@ namespace Housing.Forecast.Testing.Library
             init();
             using(_context = new ForecastContext(options)) {
                 // Arrange
-                IEnumerable<Room> rooms;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                IEnumerable<User> users;
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
                 _context.SaveChanges();
-                Room room = _context.Rooms.FirstOrDefault();
+                User user = _context.Users.FirstOrDefault();
 
                 // Act
-                rooms = _roomRepository.GetByDate(DateTime.MinValue);
+                users = _userRepository.GetByDate(DateTime.MinValue);
 
                 // Assert
-                Assert.Empty(rooms);
+                Assert.Empty(users);
             }
         }
 
@@ -82,17 +82,17 @@ namespace Housing.Forecast.Testing.Library
             using(_context = new ForecastContext(options)) {
                 // Arrange
                 IEnumerable<String> locations;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
 
-                Room room = getTestRoom();
-                room.Location = "Tampa";
-                _context.Rooms.Add(room);
+                User user = getTestUser();
+                user.Location = "Tampa";
+                _context.Users.Add(user);
 
                 _context.SaveChanges();
 
                 // Act
-                locations = _roomRepository.GetLocations();
+                locations = _userRepository.GetLocations();
 
                 // Assert
                 Assert.Equal(2, locations.Count());
@@ -104,16 +104,16 @@ namespace Housing.Forecast.Testing.Library
             init();
             using(_context = new ForecastContext(options)) {
                 // Arrange
-                IEnumerable<Room> rooms;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                IEnumerable<User> users;
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
                 _context.SaveChanges();
 
                 // Act
-                rooms = _roomRepository.GetBetweenDates(DateTime.Now, DateTime.Now);
+                users = _userRepository.GetBetweenDates(DateTime.Now, DateTime.Now);
 
                 // Assert
-                Assert.NotEmpty(rooms);
+                Assert.NotEmpty(users);
             }
         }
 
@@ -122,16 +122,16 @@ namespace Housing.Forecast.Testing.Library
             init();
             using(_context = new ForecastContext(options)) {
                 // Arrange
-                IEnumerable<Room> rooms;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                IEnumerable<User> users;
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
                 _context.SaveChanges();
 
                 // Act
-                rooms = _roomRepository.GetBetweenDates(DateTime.MinValue, DateTime.MinValue);
+                users = _userRepository.GetBetweenDates(DateTime.MinValue, DateTime.MinValue);
 
                 // Assert
-                Assert.Empty(rooms);
+                Assert.Empty(users);
             }
         }
 
@@ -140,16 +140,16 @@ namespace Housing.Forecast.Testing.Library
             init();
             using(_context = new ForecastContext(options)) {
                 // Arrange
-                IEnumerable<Room> rooms;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                IEnumerable<User> users;
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
                 _context.SaveChanges();
 
                 // Act
-                rooms = _roomRepository.GetBetweenDatesAtLocation(DateTime.Now, DateTime.Now, "Reston");
+                users = _userRepository.GetBetweenDatesAtLocation(DateTime.Now, DateTime.Now, "Reston");
 
                 // Assert
-                Assert.NotEmpty(rooms);
+                Assert.NotEmpty(users);
             }
         }
 
@@ -158,16 +158,16 @@ namespace Housing.Forecast.Testing.Library
             init();
             using(_context = new ForecastContext(options)) {
                 // Arrange
-                IEnumerable<Room> rooms;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                IEnumerable<User> users;
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
                 _context.SaveChanges();
 
                 // Act
-                rooms = _roomRepository.GetBetweenDatesAtLocation(DateTime.MinValue, DateTime.MinValue, "Reston");
+                users = _userRepository.GetBetweenDatesAtLocation(DateTime.MinValue, DateTime.MinValue, "Reston");
 
                 // Assert
-                Assert.Empty(rooms);
+                Assert.Empty(users);
             }
         }
 
@@ -176,16 +176,16 @@ namespace Housing.Forecast.Testing.Library
             init();
             using(_context = new ForecastContext(options)) {
                 // Arrange
-                IEnumerable<Room> rooms;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                IEnumerable<User> users;
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
                 _context.SaveChanges();
 
                 // Act
-                rooms = _roomRepository.GetBetweenDatesAtLocation(DateTime.Now, DateTime.Now, "test");
+                users = _userRepository.GetBetweenDatesAtLocation(DateTime.Now, DateTime.Now, "test");
 
                 // Assert
-                Assert.Empty(rooms);
+                Assert.Empty(users);
             }
         }
 
@@ -194,16 +194,16 @@ namespace Housing.Forecast.Testing.Library
             init();
             using(_context = new ForecastContext(options)) {
                 // Arrange
-                IEnumerable<Room> rooms;
-                _roomRepository = new RoomRepo(_context);
-                _context.Rooms.Add(getTestRoom());
+                IEnumerable<User> users;
+                _userRepository = new UserRepo(_context);
+                _context.Users.Add(getTestUser());
                 _context.SaveChanges();
 
                 // Act
-                rooms = _roomRepository.GetBetweenDatesAtLocation(DateTime.MinValue, DateTime.MinValue, "test");
+                users = _userRepository.GetBetweenDatesAtLocation(DateTime.MinValue, DateTime.MinValue, "test");
 
                 // Assert
-                Assert.Empty(rooms);
+                Assert.Empty(users);
             }
         }
 
@@ -213,25 +213,19 @@ namespace Housing.Forecast.Testing.Library
                 .Options;
         }
 
-        private Room getTestRoom() {
-            Room result = new Room{
-                Id = Guid.NewGuid(),
-                RoomId = Guid.NewGuid(),
-                Location = "Reston",
-                Vacancy = 1,
-                Occupancy = 1,
-                Gender = "F",
-                Address = new Address{
-                    Id = Guid.NewGuid(),
-                    AddressId = Guid.NewGuid(),
-                    Address1 = "1600 Pennsylvania Ave",
-                    Address2 = "Apt. 110-B",
-                    City = "Reston",
-                    State = "VA",
-                    PostalCode = "12345-1234",
-                    Country = "USA",
-                    Created = DateTime.Now
+        private User getTestUser() {
+            User result = new User{
+                Name = new Name {
+                    NameId = Guid.NewGuid(),
+                    First = "first",
+                    Last = "last"
                 },
+                Id = Guid.NewGuid(),
+                Location = "Reston",
+                Email = "test@test.com",
+                Gender = "M",
+                Type = "test",
+                UserId = Guid.NewGuid(),
                 Created = DateTime.Now,
                 Deleted = DateTime.MaxValue
             };
